Fixing test and removing sekvensnummer buffering
The static shutdown flag caused problems inbetween tests (because once set to `true`, it would remain `true`).
Changed to an atomic boolean that is passed to the callback, so the static is no longer needed.

Also removed buffering of sekvensnummer writes. This is not the responsibility of the callback, and should
be implemented in the SekvensnummerWriter instead.