Added work-around for bug in Skatteetatens hendelsesliste. The hendelsesliste accept two parameters, `fraSekvensnummer` and `antall`. The latter is specified as how many hendelser we want per page. This would suggest that receiving a list of fewer hendelser than `antall` indicates that the end of the list has been reached. Actually, Skatteetaten returns a list of hendelser with `fraSekvensnummer >= sekvensnummer <= (fraSekvensnummer + antall)` which means that even though we get an empty list in return there may exist hendelser with a larger sekvensnummer.
To work around this issue we use the `Første sekvensnummer fra dato` endpoint (https://skatteetaten.github.io/datasamarbeid-api-dokumentasjon/reference_feed#f%C3%B8rste-sekvensnummer-fra-dato) which gives the next sekvensnummer that would be valid on that date. It does, however, not mean that there exist a hendelse with that sekvensnummer.
If we receive an empty hendelsesliste in return, and our current `nextSekvensnummer` is smaller than `latestSekvensnummerByDato`, we set:
`nextSekvensnummer = nextSekvensnummer + antall + 1` to go to the next block of hendelser.